Fix search with tabs
